// packages/destination-actions/src/destinations/pinterest-conversions/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Settings {
  /**
   * Unique identifier of an ad account. This can be found in the Pinterest UI by following the steps mentioned [here](https://developers.pinterest.com/docs/conversions/conversion-management/#Finding%20your%20%2Cad_account_id).
   */
  ad_account_id: string
  /**
   * The conversion token for your Pinterest account. This can be found in the Pinterest UI by following the steps mentioned [here](https://developers.pinterest.com/docs/conversions/conversion-management/#Authenticating%20for%20the%20send%20conversion%20events%20endpoint).
   */
  conversion_token: string
}